--- 1l_Anvarkhujaev_2.playground/contents.xcplayground
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='5.0' target-platform='macos' buildActiveScheme='true'>
-    <timeline fileName='timeline.xctimeline'/>
-</playground>
